Fixes after review and refactor.
